workspace/SRM 673/BearSlowlySorts.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Numerics;
using Debug = System.Diagnostics.Debug;
using StringBuilder = System.Text.StringBuilder;
public class BearSlowlySorts
{
    public int minMoves(int[] w)
    {
        var min = w.Min();
        var max = w.Max();
        var a = w.OrderBy(x => x).ToArray();
        for (int i = 0; i < w.Length; i++)
            if (a[i] != w[i])
            {
                if (w[0] == min) return 1;
                if (w.Last() == max) return 1;
                if (w[0] == max && w.Last() == min) return 3;
                return 2;
            }

        return 0;
    }

    static public T[] Enumerate<T>(int n, Func<int, T> f) { var a = new T[n]; for (int i = 0; i < n; ++i) a[i] = f(i); return a; }
    static public void Swap<T>(ref T a, ref T b) { var tmp = a; a = b; b = tmp; }
}
static public class EnumerableEX
{
    static public string AsString(this IEnumerable<char> e) { return new string(e.ToArray()); }
    static public string AsJoinedString<T>(this IEnumerable<T> e, string s = " ") { return string.Join(s, e); }
}
// CUT begin
public class Tester : AbstractTester
{
    static public T[] Enumerate<T>(int n, Func<int, T> f) { var a = new T[n]; for (int i = 0; i < n; ++i) a[i] = f(i); return a; }
    public Random rand = new Random(0);
    public Tester()
    {
    }
}
// CUT end
